Documentación del código y anexo del PDF de la actividad 2 con los diagramas correspondientes.
