+++ poo/src/main/java/edu/jose/vazquez/actividades/actividad2/Starter.java
@@ -2,6 +2,7 @@ package edu.jose.vazquez.actividades.actividad2;
 
 import edu.jose.vazquez.actividades.actividad2.ui.CLI;
 
+//Clase que inicia la aplicación
 public class Starter {
     public static void main(String[] args) {
         CLI.runApp();
