+++ poo/src/main/java/edu/jose/vazquez/actividades/actividad2/models/Number.java
@@ -1,6 +1,6 @@
 package edu.jose.vazquez.actividades.actividad2.models;
 
-
+//Clase que genera los números pares del 2 al 100 y los almacena en un arreglo de enteros
 public class Number {
 
     public static int[] generarNumerosPares() {
@@ -15,16 +15,6 @@ public class Number {
         return numerosPares;
     }
 
-    public static void mostrarNumeros(int[] numeros) {
-        System.out.println("╔═════════════════════════════╗");
-        System.out.println("║ Números pares del 2 al 100: ║");
-        System.out.println("╚═════════════════════════════╝");
-        for (int i=0; i< numeros.length; i++){ 
-            System.out.printf("%6d",numeros[i]);
-            if((i+1)%10==0){
-                System.out.println();
-            }
-        }
-    }
+    
 
 }
